requirements.txt
pydub==0.25.1
spleeter==2.4.0
librosa==0.9.2
numpy==1.23.5
matplotlib==3.10.0
torch==2.6.0
torchaudio==2.6.0
soundfile==0.13.1
scipy==1.15.1
nltk==3.9.1
nemo-toolkit==2.1.0
openai==0.28.0
better-profanity==0.7.0
whisper==20240930